refactor(liblili2json): move get_course_data
